requirements.txt
deepgram-sdk==3.4.0
groq==0.9.0
langchain==0.2.12
langchain-core==0.2.28
langchain-groq==0.1.9
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.98
PyAudio==0.2.14
requests==2.32.3
